Update vcpkg baseline for HDF5 1.14.4.3 (#997)

+++ src/simplnx/Utilities/Parsing/HDF5/H5.hpp
@@ -12,7 +12,7 @@ namespace nx::core::HDF5
 {
 using IdType = int64;
 using ErrorType = int32;
-using SizeType = unsigned long long;
+using SizeType = uint64;
 
 enum class Type
 {
